intervals-icu-bot/requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.9.0
dateparser==1.2.0
